some changes in domain classes; Store and RandomStorePopulator classes created but not completely implemented;

+++ domain/src/main/java/products/Product.java
@@ -5,7 +5,7 @@ public class Product {
     private int rate;
     private double price;
 
-    public Product(String name, int rate, int price) {
+    public Product(String name, int rate, double price) {
         this.name = name;
         this.rate = rate;
         this.price = price;
